vm: rename vm test case file to `vm.rs`
